WIP: Refactor project structure
