--- src/smart-add.js
@@ -1,22 +1,0 @@
-const STATUS_MATCH = /\s*%([^\s]+)\s*/;
-const TAG_MATCH = /\s*\+([^\s]+)\s*/;
-
-const identity = v => v;
-
-const findMatch = (str, regex, modify = identity) =>
-  modify((str.match(regex) || [])[1]);
-
-const omitMatch = (str, ...regexes) =>
-  regexes.reduce((acc, curr) => acc.replace(curr, ' '), str).trim();
-
-const smartAdd = text => {
-  const description = omitMatch(text, TAG_MATCH, STATUS_MATCH);
-  const status = findMatch(text, STATUS_MATCH, v =>
-    ['done', 'in-progress', 'todo'].includes(v) ? v : 'todo',
-  );
-  const tag = findMatch(text, TAG_MATCH);
-
-  return { description, status, tag };
-};
-
-module.exports = smartAdd;
